Encapsulated the caching functionality is a separate file.
Files checked in: DataRepo/models.py DataRepo/utils.py DataRepo/hier_cached_model.py
